+++ lib/src/endpoints/api/api_endpoint.dart
@@ -1,14 +1,14 @@
 import 'dart:async';
 
+import 'package:uwdart/src/base/client.dart';
+import 'package:uwdart/src/base/endpoints.dart';
+import 'package:uwdart/src/base/uw_endpoint.dart';
+import 'package:uwdart/src/data/api_request.dart';
 import 'package:uwdart/src/endpoints/api/model/api_changelog.dart';
 import 'package:uwdart/src/endpoints/api/model/api_method.dart';
 import 'package:uwdart/src/endpoints/api/model/api_service.dart';
 import 'package:uwdart/src/endpoints/api/model/api_usage.dart';
 import 'package:uwdart/src/endpoints/api/model/api_version.dart';
-import 'package:uwdart/src/base/client.dart';
-import 'package:uwdart/src/base/endpoints.dart';
-import 'package:uwdart/src/base/uw_endpoint.dart';
-import 'package:uwdart/src/data/api_request.dart';
 
 /// ApiEndpoint provides a wrapper around the raw Api Endpoints from UW Open Data API
 /// https://github.com/uWaterloo/api-documentation#api