+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.5
 ### Added
 - Added browser and io version of the library
-- Added Food Services Endpoint
+- Added Food Services, Feds, and Awards Endpoints
 
 ### Modified
 - Generic open, close and makeRequest functions removed